Fix memory leaks and change the generic quicksort to not allocate extra memory.
All the sort-tester functions for the different cases made copies of the data in order
to use the same raw data for every algorithm. However, the function did not free the
memory allocated for that copy before it ended, making the pointer go out of scope and
the memory unreachable. Also, the timeval struct used for measuring the elapsed time for
each algorithm is allocated on the heap by the create_start_point() and was not being
released by the function, resulting in another memory leak.

The generic quicksort function was allocating an element's size of memory on every call
to copy the pivot but it was not freeing it, resulting in a memory leak. However, the copy
is unecessary and so the heap allocation was removed and replaced with a simple reference
to the memory position of the pivot.